refactor(leetcode/.../roman_to_integer)
Made a module
